pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graph2svg"
version = "0.1.0"
description = "Convert hand-drawn graph drafts to publication-quality vector graphics"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.8",
    "networkx>=3.0",
    "pydantic>=2.0",
    "openai>=1.0",
    "Pillow>=10.0",
    "click>=8.0",
]

[project.scripts]
graph2svg = "graph2svg.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["graph2svg*"]
